created register-user route, encrypts password
