+++ package.json
@@ -16,6 +16,7 @@
   },
   "homepage": "https://github.com/macdhollister/skill.it#readme",
   "dependencies": {
+    "bcrypt-nodejs": "0.0.3",
     "dotenv": "^6.2.0",
     "ejs": "^2.6.1",
     "express": "^4.16.3",
