+++ server.js
@@ -11,6 +11,9 @@ app.use(express.static("public"));
 app.set("view engine", "ejs");
 
 // Routes
+const authRoutes = require("./routes/authRoutes");
+app.use(authRoutes);
+
 const dataRoutes = require("./routes/dataRoutes");
 app.use(dataRoutes);
